neuroformer_project/requirements.txt
# Core dependencies
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
numpy==1.23.5
scipy==1.9.3
pandas==1.5.3

# Deep learning utilities
pytorch-lightning==1.8.6
tensorboard==2.11.2
wandb==0.13.10

# EEG processing
mne==1.3.0
pywavelets==1.4.1
scipy==1.9.3

# Eye-tracking processing
pymovements==0.4.0
scikit-learn==1.2.1

# Data handling
h5py==3.8.0
tables==3.8.0
pyarrow==11.0.0

# Visualization
matplotlib==3.6.3
seaborn==0.12.2
plotly==5.13.0
pillow==9.4.0

# Metrics and evaluation
scikit-learn==1.2.1
scipy==1.9.3
statsmodels==0.13.5

# Configuration management
pyyaml==6.0
omegaconf==2.3.0
hydra-core==1.3.1

# Utilities
tqdm==4.64.1
joblib==1.2.0
filelock==3.9.0

# Testing
pytest==7.2.1
pytest-cov==4.0.0

# Code quality
black==23.1.0
flake8==6.0.0
mypy==1.0.0
isort==5.12.0

# Jupyter notebooks
jupyter==1.0.0
ipykernel==6.21.1
ipywidgets==8.0.4

# Optional: GPU monitoring
gpustat==1.0.0
nvidia-ml-py3==7.352.0

# Optional: Distributed training
torch-distributed==0.1.0
horovod==0.27.0

# Documentation
sphinx==5.3.0
sphinx-rtd-theme==1.2.0
